+++ Changes.md
@@ -24,7 +24,9 @@ Fixes
 -----
 
 - GraphEditor : Implemented undo for node drags (#423).
-- Reference : Fixed bug which caused promoted Spreadsheet values to be lost when exporting a Box for referencing.
+- Reference :
+  - Fixed bug which caused promoted Spreadsheet values to be lost when exporting a Box for referencing.
+  - Fixed bug which caused promoted Spreadsheet cells to be duplicated by copy/paste.
 - DeleteSets : Fixed bug which allowed the deletion of Gaffer's internal `__lights`, `__cameras` and `__lightFilters` sets. These are now always preserved, because they are needed to output the scene for rendering.
 - ImageReader/ImageWriter : Fixed handling of errors in Python functions registered using `setDefaultColorSpaceFunction()`.
 - StyleSheet : Fixed monospace font stack.